+++ workflow/Snakefile
@@ -15,14 +15,13 @@ report: "report/workflow.rst"
 container: "docker://continuumio/miniconda3"
 
 
-# rule all:
-#     input:
+include: "rules/common.smk"
+include: "rules/extract-data.smk"
+include: "rules/process-data.smk"
+
+
 # The first rule should define the default target files
 # Subsequent target rules can be specified below. They should start with all_*.
-
-
-include: "rules/common.smk"
-include: "rules/other.smk"
-include: "rules/extract-data.smk"
-include: "rules/scan_decomp_data.smk"
-include: "rules/process-data.smk"
+rule all:
+    input:
+        expand("results/done/{id}", id=get_all_ids()),
