--- workflow/envs/python.yaml
@@ -1,7 +1,0 @@
-channels:
-  - conda-forge
-  - bioconda
-  - anaconda
-dependencies:
-  - python =3.9.2
-
