--- workflow/rules/other.smk
@@ -1,1 +1,0 @@
-# An example collection of Snakemake rules imported in the main Snakefile.
