+++ workflow/scripts/redact-page.py
@@ -1,20 +1,10 @@
 import typing
-import json
 
 import cv2
-import Levenshtein
 import pandas as pd
-import pytesseract
-from pytesseract import Output
 
 
-def parse_page(
-    image_path: str,
-    out_path: str,
-    personal_data_path: dict,
-    min_conf: float = 0.6,
-    max_dist: int = 2,
-):
+def process_page(image_path: str, out_path: str, data_to_redact: str):
     """Analyzes the passed image and removes personal information.
 
     Args:
@@ -25,108 +15,15 @@ def parse_page(
         max_dist (int, optional): maximum Levenshtein distance of the found text on the image to the personal data. Defaults to 2.
     """
 
+    df = pd.read_csv(data_to_redact)
     img = cv2.imread(image_path)
 
-    with open(personal_data_path) as json_file:
-        personal_data = json.load(json_file)
-
-    df = detect_text(img, min_conf)
-    df = select_personal_data(df, personal_data, max_dist)
     img = redact(df, img)
 
+    if not ".jpg" in out_path[-3:]:
+        "".join([out_path, ".jpg"])
+
     cv2.imwrite(out_path, img)
-
-
-def detect_text(img: typing.Any, min_conf: float) -> pd.DataFrame:
-    """Recognizes text on the image.
-
-    Args:
-        img (typing.Any): image with text to be recognized.
-        min_conf (float): minimum OCR Confidence Scores.
-
-    Returns:
-        pd.DataFrame: all found text on image with text field data, filtered by min_conf.
-    """
-
-    # ocr
-    detected_text_df = pytesseract.image_to_data(img, output_type=Output.DATAFRAME)
-
-    # filter ocr table
-    detected_text_df = detected_text_df[detected_text_df.conf >= min_conf]
-    detected_text_df.drop(
-        columns=["level", "page_num", "block_num", "par_num", "line_num", "word_num"],
-        inplace=True,
-    )
-    detected_text_df.text = detected_text_df.text.str.lower()
-
-    return detected_text_df
-
-
-def select_personal_data(
-    detected_text_df: pd.DataFrame, personal_data: dict, min_distance: int
-) -> pd.DataFrame:
-    """Identifies personal data from the detected text.
-
-    Args:
-        detected_text_df (pd.DataFrame): detected text on the image.
-        personal_data (dict): personal data to be masked out
-        min_distance (int): maximum Levenshtein distance of the found text on the image to the personal data.
-
-    Returns:
-        pd.DataFrame: person data with location on image, filtered by min_distance.
-    """
-
-    final_df = pd.DataFrame()
-
-    max_spaces = max([value.count(" ") for value in personal_data.values()])
-    for no_spaces in range(max_spaces + 1):
-        tmp_df = detected_text_df.copy()
-        tmp_df.rename(columns={"text": "text_0", "width": "width_0"}, inplace=True)
-
-        # subset of personal data dict, according to # spaces
-        tmp_dict = {
-            key: value.replace(" ", "")
-            for key, value in personal_data.items()
-            if value.count(" ") == no_spaces
-        }
-
-        # shif text to get longer phrases
-        if no_spaces > 0:
-            for shift in range(1, no_spaces + 1):
-                # shift text column and aggreagte
-                highest_text_column_name = "text_" + str(shift)
-                tmp_df[highest_text_column_name] = tmp_df[
-                    "text_" + str(shift - 1)
-                ] + tmp_df.text_0.shift(-shift).fillna("")
-
-                # shift width column and aggreagte
-                highest_width_column_name = "width_" + str(shift)
-                tmp_df[highest_width_column_name] = (
-                    tmp_df["width_" + str(shift - 1)]
-                    + tmp_df.width_0.shift(-shift).fillna(0)
-                    + tmp_df.width_0 / tmp_df.text_0.str.len()
-                )
-
-            tmp_df["width_0"] = tmp_df[highest_width_column_name].astype(int)
-        else:
-            highest_text_column_name = "text_0"
-
-        # calc edit distances for each key in subsampled dict
-        for key, value in tmp_dict.items():
-            tmp_df[key] = tmp_df[highest_text_column_name].apply(
-                lambda text: Levenshtein.distance(value, text)
-            )
-
-        tmp_df.rename(columns={"text_0": "text", "width_0": "width"}, inplace=True)
-        # select entries, where distance is below or equal a threshhold
-        query_list = []
-        for col in tmp_dict.keys():
-            query_list.append(f"{col}<={min_distance}")
-        tmp_df = tmp_df.query(" | ".join(query_list))
-
-        final_df = final_df.append(tmp_df, ignore_index=True)
-
-    return final_df[["left", "top", "width", "height", "conf", "text"]]
 
 
 def redact(personal_data_df: pd.DataFrame, img: typing.Any) -> typing.Any:
@@ -139,7 +36,7 @@ def redact(personal_data_df: pd.DataFrame, img: typing.Any) -> typing.Any:
     Returns:
         typing.Any: redacted image.
     """
-    
+
     for i in personal_data_df.index:
         (x, y, w, h) = (
             personal_data_df.loc[i].left,
@@ -152,7 +49,9 @@ def redact(personal_data_df: pd.DataFrame, img: typing.Any) -> typing.Any:
 
 
 if __name__ == "__main__":
-
     sys.stderr = open(snakemake.log[0], "w")
-
-    parse_page(image_path = snakemake.input.orginal_img, out_path = snakemake.output[0], personal_data_path = snakemake.input.personal_data)
+    process_page(
+        image_path=snakemake.input.orginal_page,
+        out_path=snakemake.output[0],
+        data_to_redact=snakemake.input.data_to_redact,
+    )
